+++ packages/client-connectivity/RELEASES.md
@@ -109,3 +109,13 @@ Based on:
 - [typescript v1.0.0-alpha.10] packages/client-connectivity
 ### Releases
 - [NPM v1.0.0-alpha.10] https://www.npmjs.com/package/@formance/sdk-connectivity/v/1.0.0-alpha.10 - packages/client-connectivity
+
+## 2024-07-31 00:18:10
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.349.0 (2.382.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.0.0-alpha.11] packages/client-connectivity
+### Releases
+- [NPM v1.0.0-alpha.11] https://www.npmjs.com/package/@formance/sdk-connectivity/v/1.0.0-alpha.11 - packages/client-connectivity